InputCommon: fix some compilation and Lint warnings

+++ Source/Core/Core/FreeLookManager.cpp
@@ -244,9 +244,9 @@ ControllerEmu::ControlGroup* GetInputGroup(int pad_num, FreeLookGroup group)
 
 void UpdateInput()
 {
-  for (int i = 0; i < s_config.GetControllerCount(); i++)
+  for (size_t i = 0; i < s_config.GetControllerCount(); i++)
   {
-    static_cast<FreeLookController*>(s_config.GetController(i))->Update();
+    static_cast<FreeLookController*>(s_config.GetController(int(i)))->Update();
   }
 }
